Replace old adapters with new ones
