+++ app/src/main/java/com/firebase/uidemo/database/ChatActivity.java
@@ -148,7 +148,7 @@ public class ChatActivity extends AppCompatActivity implements FirebaseAuth.Auth
             }
 
             @Override
-            protected void onDataChanged() {
+            public void onDataChanged() {
                 // If there are no chat messages, show a view that invites the user to add a message.
                 mEmptyListMessage.setVisibility(mAdapter.getItemCount() == 0 ? View.VISIBLE : View.GONE);
             }
