+++ database/src/main/java/com/firebase/ui/database/FirebaseListAdapter.java
@@ -1,17 +1,3 @@
-/*
- * Copyright 2016 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.firebase.ui.database;
 
 import android.app.Activity;
@@ -20,6 +6,7 @@ import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
+import android.widget.ListView;
 
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
@@ -27,91 +14,82 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.Query;
 
 /**
- * This class is a generic way of backing an Android ListView with a Firebase location.
- * It handles all of the child events at the given Firebase location. It marshals received data into the given
- * class type. Extend this class and provide an implementation of {@code populateView}, which will be given an
- * instance of your list item mLayout and an instance your class that holds your data.
- * Simply populate the view however you like and this class will handle updating the list as the data changes.
+ * This class is a generic way of backing an Android {@link android.widget.ListView} with a Firebase
+ * location. It handles all of the child events at the given Firebase location. It marshals received
+ * data into the given class type.
  * <p>
- * <pre>
- *     DatabaseReference ref = FirebaseDatabase.getInstance().getReference();
- *     ListAdapter adapter = new FirebaseListAdapter<ChatMessage>(
- *              this, ChatMessage.class, android.R.layout.two_line_list_item, ref)
- *     {
- *         protected void populateView(View view, ChatMessage chatMessage, int position)
- *         {
- *             ((TextView)view.findViewById(android.R.id.text1)).setText(chatMessage.getName());
- *             ((TextView)view.findViewById(android.R.id.text2)).setText(chatMessage.getMessage());
- *         }
- *     };
- *     listView.setListAdapter(adapter);
- * </pre>
+ * See the <a href="https://github.com/firebase/FirebaseUI-Android/blob/master/database/README.md">README</a>
+ * for an in-depth tutorial on how to set up the FirebaseListAdapter.
  *
  * @param <T> The class type to use as a model for the data contained in the children of the given
  *            Firebase location
- * @deprecated use {@link com.firebase.ui.database.adapter.FirebaseRecyclerAdapter} instead
  */
-@Deprecated
-public abstract class FirebaseListAdapter<T> extends BaseAdapter {
+public abstract class FirebaseListAdapter<T> extends BaseAdapter implements FirebaseAdapter<T> {
     private static final String TAG = "FirebaseListAdapter";
 
-    private FirebaseArray mSnapshots;
-    private final Class<T> mModelClass;
     protected Activity mActivity;
+    protected FirebaseArray mSnapshots;
+    protected Class<T> mModelClass;
     protected int mLayout;
-    private ChangeEventListener mListener;
 
-    FirebaseListAdapter(Activity activity,
-                        Class<T> modelClass,
-                        @LayoutRes int modelLayout,
-                        FirebaseArray snapshots) {
+    /**
+     * @param activity    The {@link Activity} containing the {@link ListView}
+     * @param modelClass  Firebase will marshall the data at a location into an instance of a class
+     *                    that you provide
+     * @param modelLayout This is the layout used to represent a single list item. You will be
+     *                    responsible for populating an instance of the corresponding view with the
+     *                    data from an instance of modelClass.
+     * @param snapshots   The data used to populate the adapter
+     */
+    public FirebaseListAdapter(Activity activity,
+                               FirebaseArray snapshots,
+                               Class<T> modelClass,
+                               @LayoutRes int modelLayout) {
         mActivity = activity;
+        mSnapshots = snapshots;
         mModelClass = modelClass;
         mLayout = modelLayout;
-        mSnapshots = snapshots;
 
-        mListener = mSnapshots.addChangeEventListener(new ChangeEventListener() {
-            @Override
-            public void onChildChanged(EventType type, int index, int oldIndex) {
-                FirebaseListAdapter.this.onChildChanged(type, index, oldIndex);
-            }
-
-            @Override
-            public void onDataChanged() {
-                FirebaseListAdapter.this.onDataChanged();
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                FirebaseListAdapter.this.onCancelled(error);
-            }
-        });
+        startListening();
     }
 
     /**
-     * @param activity    The activity containing the ListView
-     * @param modelClass  Firebase will marshall the data at a location into
-     *                    an instance of a class that you provide
-     * @param modelLayout This is the layout used to represent a single list item.
-     *                    You will be responsible for populating an instance of the corresponding
-     *                    view with the data from an instance of modelClass.
-     * @param ref         The Firebase location to watch for data changes. Can also be a slice of a location,
-     *                    using some combination of {@code limit()}, {@code startAt()}, and {@code endAt()}.
+     * @param query The Firebase location to watch for data changes. Can also be a slice of a
+     *              location, using some combination of {@code limit()}, {@code startAt()}, and
+     *              {@code endAt()}.
+     * @see #FirebaseListAdapter(Activity, FirebaseArray, Class, int)
      */
     public FirebaseListAdapter(Activity activity,
                                Class<T> modelClass,
-                               int modelLayout,
-                               Query ref) {
-        this(activity, modelClass, modelLayout, new FirebaseArray(ref));
-    }
-
-    public void cleanup() {
-        mSnapshots.removeChangeEventListener(mListener);
+                               @LayoutRes int modelLayout,
+                               Query query) {
+        this(activity, new FirebaseArray(query), modelClass, modelLayout);
     }
 
     @Override
-    public int getCount() {
-        return mSnapshots.size();
+    public void startListening() {
+        if (!mSnapshots.isListening(this)) {
+            mSnapshots.addChangeEventListener(this);
+        }
+    }
+
+    @Override
+    public void cleanup() {
+        mSnapshots.removeChangeEventListener(this);
+    }
+
+    @Override
+    public void onChildChanged(ChangeEventListener.EventType type, int index, int oldIndex) {
+        notifyDataSetChanged();
+    }
+
+    @Override
+    public void onDataChanged() {
+    }
+
+    @Override
+    public void onCancelled(DatabaseError error) {
+        Log.w(TAG, error.toException());
     }
 
     @Override
@@ -119,19 +97,19 @@ public abstract class FirebaseListAdapter<T> extends BaseAdapter {
         return parseSnapshot(mSnapshots.get(position));
     }
 
-    /**
-     * This method parses the DataSnapshot into the requested type. You can override it in subclasses
-     * to do custom parsing.
-     *
-     * @param snapshot the DataSnapshot to extract the model from
-     * @return the model extracted from the DataSnapshot
-     */
-    protected T parseSnapshot(DataSnapshot snapshot) {
+    @Override
+    public T parseSnapshot(DataSnapshot snapshot) {
         return snapshot.getValue(mModelClass);
     }
 
+    @Override
     public DatabaseReference getRef(int position) {
         return mSnapshots.get(position).getRef();
+    }
+
+    @Override
+    public int getCount() {
+        return mSnapshots.size();
     }
 
     @Override
@@ -151,26 +129,6 @@ public abstract class FirebaseListAdapter<T> extends BaseAdapter {
         // Call out to subclass to marshall this model into the provided view
         populateView(view, model, position);
         return view;
-    }
-
-    /**
-     * @see ChangeEventListener#onChildChanged(ChangeEventListener.EventType, int, int)
-     */
-    protected void onChildChanged(ChangeEventListener.EventType type, int index, int oldIndex) {
-        notifyDataSetChanged();
-    }
-
-    /**
-     * @see ChangeEventListener#onDataChanged()
-     */
-    protected void onDataChanged() {
-    }
-
-    /**
-     * @see ChangeEventListener#onCancelled(DatabaseError)
-     */
-    protected void onCancelled(DatabaseError error) {
-        Log.w(TAG, error.toException());
     }
 
     /**
